+++ monolith/src/main/java/com/amazon/aws/partners/saasfactory/service/ProductService.java
@@ -8,7 +8,7 @@ public interface ProductService {
 
     public List<Product> getProducts() throws Exception;
 
-    public Product getProduct(Long productId) throws Exception;
+    public Product getProduct(Integer productId) throws Exception;
 
     public Product saveProduct(Product product) throws Exception;
